move getAtomsRegistry tests. add callback registry to SitecoreProvider
